Add parameter and use quiet mode of GDAL

+++ create_contours.py
@@ -21,15 +21,15 @@ def smoothTerrain(inputTerrain, kernelSize):
     prefix = randomString()
     smooth = kernelSize
 
-    os.system("gdal_translate -ot Float32 -a_nodata -32768 "  + inputDEM + " " + prefix + "_dem.tif")
+    os.system("gdal_translate -ot Float32 -a_nodata -32768 "  + inputDEM + " " + prefix + "_dem.tif -q")
 
     # Calculate TPI
     print("Calculate TPI...")
-    os.system("gdaldem TPI " + inputDEM + " " + prefix + "_dem_tpi.tif")
+    os.system("gdaldem TPI " + inputDEM + " " + prefix + "_dem_tpi.tif -q")
 
     # Build VRT for soft smoothed DEM
     print("Smooth DEM...")
-    os.system("gdalbuildvrt " + prefix + "_dem_blur_3x3.vrt " + prefix + "_dem.tif")
+    os.system("gdalbuildvrt " + prefix + "_dem_blur_3x3.vrt " + prefix + "_dem.tif -q")
 
     file = open(prefix + "_dem_blur_3x3.vrt", "rt")
     data = file.read()
@@ -42,7 +42,7 @@ def smoothTerrain(inputTerrain, kernelSize):
     file.close()
 
     # Build VRT for smoothed DEM
-    os.system("gdalbuildvrt " + prefix + "_dem_blur_5x5.vrt " + prefix + "_dem.tif")
+    os.system("gdalbuildvrt " + prefix + "_dem_blur_5x5.vrt " + prefix + "_dem.tif -q")
 
     file = open(prefix + "_dem_blur_5x5.vrt", "rt")
     data = file.read()
@@ -55,7 +55,7 @@ def smoothTerrain(inputTerrain, kernelSize):
     file.close()
 
     # Build VRT for more smoothed DEM
-    os.system("gdalbuildvrt " + prefix + "_dem_blur_7x7.vrt " + prefix + "_dem.tif")
+    os.system("gdalbuildvrt " + prefix + "_dem_blur_7x7.vrt " + prefix + "_dem.tif -q")
 
     file = open(prefix + "_dem_blur_7x7.vrt", "rt")
     data = file.read()
@@ -68,7 +68,7 @@ def smoothTerrain(inputTerrain, kernelSize):
     file.close()
 
     # Build VRT for more stronger smoothed DEM
-    os.system("gdalbuildvrt " + prefix + "_dem_blur_9x9.vrt " + prefix + "_dem.tif")
+    os.system("gdalbuildvrt " + prefix + "_dem_blur_9x9.vrt " + prefix + "_dem.tif -q")
 
     file = open(prefix + "_dem_blur_9x9.vrt", "rt")
     data = file.read()
@@ -81,7 +81,7 @@ def smoothTerrain(inputTerrain, kernelSize):
     file.close()
 
     # Build VRT for strong smoothed DEM
-    os.system("gdalbuildvrt " + prefix + "_dem_blur_13x13.vrt " + prefix + "_dem.tif")
+    os.system("gdalbuildvrt " + prefix + "_dem_blur_13x13.vrt " + prefix + "_dem.tif -q")
 
     file = open(prefix + "_dem_blur_13x13.vrt", "rt")
     data = file.read()
@@ -95,9 +95,9 @@ def smoothTerrain(inputTerrain, kernelSize):
 
     print("Reclassify and smooth TPI...")
     # Reclassify TPI
-    os.system('gdal_calc.py -A ' + prefix + '_dem_tpi.tif --outfile=' + prefix + '_tpi_pos.tif --NoDataValue=-32768 --calc="((-1)*A*(A<0))+(A*(A>=0))"')
+    os.system('gdal_calc.py -A ' + prefix + '_dem_tpi.tif --outfile=' + prefix + '_tpi_pos.tif --NoDataValue=-32768 --calc="((-1)*A*(A<0))+(A*(A>=0))" --quiet')
     # Build VRT for smooth TPI
-    os.system("gdalbuildvrt " + prefix + "_tpi_blur_3x3.vrt " + prefix + "_tpi_pos.tif")
+    os.system("gdalbuildvrt " + prefix + "_tpi_blur_3x3.vrt " + prefix + "_tpi_pos.tif -q")
 
     file = open(prefix + "_tpi_blur_3x3.vrt", "rt")
     data = file.read()
@@ -114,22 +114,22 @@ def smoothTerrain(inputTerrain, kernelSize):
     # Normalize smoothed TPI, if no max avaible than just copy file
     try:
         maxValue = re.findall('[0-9]*\.[0-9]*', re.findall('STATISTICS_MAXIMUM=\d*.\d*', info)[0])[0]
-        os.system('gdal_calc.py -A ' + prefix + '_tpi_blur_3x3.vrt --outfile=' + prefix + '_tpi_norm.tif --NoDataValue=-32768 --calc="A / ' + maxValue + '"')
+        os.system('gdal_calc.py -A ' + prefix + '_tpi_blur_3x3.vrt --outfile=' + prefix + '_tpi_norm.tif --NoDataValue=-32768 --calc="A / ' + maxValue + '" --quiet')
     except:
-        os.system("gdal_translate " + prefix + "_tpi_blur_3x3.vrt " + prefix + "_tpi_norm.tif")
+        os.system("gdal_translate " + prefix + "_tpi_blur_3x3.vrt " + prefix + "_tpi_norm.tif -q")
 
     # Combine it together
     print("Build better DEM for contour lines...")
     if (smooth == 13):
-        os.system('gdal_calc.py -A ' + prefix + '_tpi_norm.tif -B ' + prefix + '_dem_blur_3x3.vrt -C ' + prefix + '_dem_blur_13x13.vrt --outfile="smooth_' + inputDEM + '" --overwrite --calc="A*B+(1-A)*C"')
+        os.system('gdal_calc.py -A ' + prefix + '_tpi_norm.tif -B ' + prefix + '_dem_blur_3x3.vrt -C ' + prefix + '_dem_blur_13x13.vrt --outfile="smooth_' + inputDEM + '" --overwrite --calc="A*B+(1-A)*C" --quiet')
     if (smooth == 7):
-        os.system('gdal_calc.py -A ' + prefix + '_tpi_norm.tif -B ' + prefix + '_dem_blur_3x3.vrt -C ' + prefix + '_dem_blur_7x7.vrt --outfile="smooth_' + inputDEM + '" --overwrite --calc="A*B+(1-A)*C"')   
+        os.system('gdal_calc.py -A ' + prefix + '_tpi_norm.tif -B ' + prefix + '_dem_blur_3x3.vrt -C ' + prefix + '_dem_blur_7x7.vrt --outfile="smooth_' + inputDEM + '" --overwrite --calc="A*B+(1-A)*C" --quiet')   
     if (smooth == 5):
-        os.system('gdal_calc.py -A ' + prefix + '_tpi_norm.tif -B ' + prefix + '_dem_blur_3x3.vrt -C ' + prefix + '_dem_blur_5x5.vrt --outfile="smooth_' + inputDEM + '" --overwrite --calc="A*B+(1-A)*C"')    
+        os.system('gdal_calc.py -A ' + prefix + '_tpi_norm.tif -B ' + prefix + '_dem_blur_3x3.vrt -C ' + prefix + '_dem_blur_5x5.vrt --outfile="smooth_' + inputDEM + '" --overwrite --calc="A*B+(1-A)*C" --quiet')    
     if (smooth == 3):
-        os.system('gdal_calc.py -A ' + prefix + '_tpi_norm.tif -B ' + prefix + '_dem_blur_3x3.vrt -C ' + prefix + '_dem_blur_3x3.vrt --outfile="smooth_' + inputDEM + '" --overwrite --calc="A*B+(1-A)*C"')   
+        os.system('gdal_calc.py -A ' + prefix + '_tpi_norm.tif -B ' + prefix + '_dem_blur_3x3.vrt -C ' + prefix + '_dem_blur_3x3.vrt --outfile="smooth_' + inputDEM + '" --overwrite --calc="A*B+(1-A)*C" --quiet')   
     else:
-        os.system('gdal_calc.py -A ' + prefix + '_tpi_norm.tif -B ' + prefix + '_dem_blur_3x3.vrt -C ' + prefix + '_dem_blur_9x9.vrt --outfile="smooth_' + inputDEM + '" --overwrite --calc="A*B+(1-A)*C"')
+        os.system('gdal_calc.py -A ' + prefix + '_tpi_norm.tif -B ' + prefix + '_dem_blur_3x3.vrt -C ' + prefix + '_dem_blur_9x9.vrt --outfile="smooth_' + inputDEM + '" --overwrite --calc="A*B+(1-A)*C" --quiet')
 
     # Clean up
     os.remove(prefix + "_dem.tif")
@@ -147,30 +147,36 @@ def smoothTerrain(inputTerrain, kernelSize):
     return("smooth_" + inputDEM)
 
 
-def createContours(outputFile, smooth_dem, interval):
-    # temp sqlite database
-    buffer = 20
-    doubleInterval = float(interval) / 2
+def createContours(outputFile, smooth_dem, interval, contourBuffer):
+    # define variables
+    buffer = contourBuffer
     prefix = randomString()
     contoursFile = prefix + "_contour.gpkg"
 
-    os.system("gdal_contour -inodata -a ele "  + smooth_dem + " " + contoursFile + " -i " + str(doubleInterval))
+    if contourBuffer != 0:
+        doubleInterval = float(interval) / 2
+    else:
+        doubleInterval = float(interval) 
+
+    os.system(f"gdal_contour -inodata -a ele  {smooth_dem} {contoursFile} -i {doubleInterval} -q ")
    
     # Calculate length of countour lines
-    print("Calculate lenght of contour lines.")
+    print("Calculate lenght of contour lines...")
     os.system(f"ogr2ogr {contoursFile} {contoursFile} -update -dialect SQLITE -sql 'ALTER TABLE contour ADD COLUMN line_length float'")
     os.system(f"ogr2ogr {contoursFile} {contoursFile} -update -dialect SQLITE -sql 'UPDATE contour SET line_length = ST_Length(geom)'")
-    # Calculate additional contour lines for flatter areas
-    print("Calculate additional contour lines for flatter areas.")
-    os.system(f"ogr2ogr {contoursFile} {contoursFile} -append -nln 'buffer' -dialect SQLITE -sql 'SELECT id, ele, line_length, ST_Union(ST_Buffer(geom, {buffer} )) as geom FROM contour WHERE (ele % {interval} = 0)'")
-    os.system(f"ogr2ogr {contoursFile} {contoursFile} -append -nln 'bbox' -dialect SQLITE -sql 'SELECT ST_Envelope(ST_Union(geom)) as geom FROM contour'")
-    os.system(f"ogr2ogr {contoursFile} {contoursFile} -append -nln 'diff' -dialect SQLITE -sql 'SELECT ST_Difference(bbox.geom, buffer.geom) as geom FROM bbox, buffer'")
-    os.system(f"ogr2ogr {contoursFile} {contoursFile} -append -nln 'clip' -nlt PROMOTE_TO_MULTI -dialect SQLITE -sql 'SELECT contour.id, contour.line_length, (ST_intersection(diff.geom, contour.geom)) AS geom FROM diff JOIN contour ON ST_intersects(contour.geom, diff.geom)'")
-    os.system(f"ogr2ogr {contoursFile} {contoursFile} -update -dialect SQLITE -sql 'UPDATE clip SET line_length = ST_Length(geom)'")
-    # Remove temp 
-    os.system(f"ogr2ogr {contoursFile} {contoursFile} -update -dialect SQL -sql 'DROP TABLE buffer'")
-    os.system(f"ogr2ogr {contoursFile} {contoursFile} -update -dialect SQL -sql 'DROP TABLE bbox'")
-    os.system(f"ogr2ogr {contoursFile} {contoursFile} -update -dialect SQL -sql 'DROP TABLE diff'")
+    
+    if contourBuffer != 0:
+        # Calculate additional contour lines for flatter areas
+        print("Calculate additional contour lines for flatter areas...")
+        os.system(f"ogr2ogr {contoursFile} {contoursFile} -append -nln 'buffer' -dialect SQLITE -sql 'SELECT id, ele, line_length, ST_Union(ST_Buffer(geom, {buffer} )) as geom FROM contour WHERE (ele % {interval} = 0)'")
+        os.system(f"ogr2ogr {contoursFile} {contoursFile} -append -nln 'bbox' -dialect SQLITE -sql 'SELECT ST_Envelope(ST_Union(geom)) as geom FROM contour'")
+        os.system(f"ogr2ogr {contoursFile} {contoursFile} -append -nln 'diff' -dialect SQLITE -sql 'SELECT ST_Difference(bbox.geom, buffer.geom) as geom FROM bbox, buffer'")
+        os.system(f"ogr2ogr {contoursFile} {contoursFile} -append -nln 'contours_cliped' -nlt PROMOTE_TO_MULTI -dialect SQLITE -sql 'SELECT contour.id, contour.line_length, (ST_intersection(diff.geom, contour.geom)) AS geom FROM diff JOIN contour ON ST_intersects(contour.geom, diff.geom)'")
+        os.system(f"ogr2ogr {contoursFile} {contoursFile} -update -dialect SQLITE -sql 'UPDATE contours_cliped SET line_length = ST_Length(geom)'")
+        # Remove temp 
+        os.system(f"ogr2ogr {contoursFile} {contoursFile} -update -dialect SQL -sql 'DROP TABLE buffer'")
+        os.system(f"ogr2ogr {contoursFile} {contoursFile} -update -dialect SQL -sql 'DROP TABLE bbox'")
+        os.system(f"ogr2ogr {contoursFile} {contoursFile} -update -dialect SQL -sql 'DROP TABLE diff'")
 
     # Convert to output format
     os.system(f"ogr2ogr {outputFile} {contoursFile}")
@@ -185,26 +191,28 @@ def main(argv):
     interval = ''
     pixelSize = ''
     gaussainBlur = ''
+    contourBuffer = 0
 
     tmp_file = "resample_dem.tif"
     overwrite = True
 
     try:
-        opts, args = getopt.getopt(argv, "", ["help", "inputDEM=", "outputFile=", "intervall=", "pixelSize=", "gaussainBlur="]) 
+        opts, args = getopt.getopt(argv, "", ["help", "inputDEM=", "outputFile=", "interval=", "pixelSize=", "gaussainBlur=", "addContoursBuffer="]) 
     except getopt.GetoptError:
         print("Please look use create_contours.py --help to get more information how to use the script. ")
         sys.exit(2)
     for opt, arg in opts:
         if opt in ("--help"):
             print("Create smooth contours with the help of GDAL. Usage:")
-            print("\tcreate_contour.py --inputDEM=DEMfile.tif --outputFile=contours.geoJSON --intervall=5 --pixelSize=2")
+            print("\tcreate_contour.py --inputDEM=DEMfile.tif --outputFile=contours.geoJSON --interval=5 --pixelSize=2")
             print("Algorithm after: P. Kettunen, C. Koski, and J. Oksanen, 'A design of contour generation for topographic maps with adaptive DEM smoothing' International Journal of Cartography, vol. 3, no. 1, pp. 19–30, Jun. 2017.")
             print("Parameters:")
             print("\t--inputDEM=(rasterFilename)* put here your DEM in any GDAL readable format.")
             print("\t--outputFile=(vectorFilename)* write the name of the outputfile for the contours. The format is guessed form the extention")
-            print("\t--intervall=(float)* intervall between the contour lines.")
+            print("\t--interval=(float)* interval between the contour lines.")
             print("\t--pixelSize=(float)* pixel size of the DEM, or a greater number. Used for reampling and should correspond with the aimed map scale")
             print("\t--gaussainBlur={3, 5, 7, 9, 13} kernel size for smoothing flat areas. Default 9.")
+            print("\t--addContoursBuffer=(float) create addional contour lines in between and cut them to the buffer. Default 0 for no execution.")
             print("\t--help print this hopefully helpfull text.")
             print("\t*necessary for execution.")
             print("Mathias Gröbe 2020")
@@ -213,17 +221,19 @@ def main(argv):
             inputDEM = arg #dem_file.tif
         elif opt in ("--outputFile"):
             outputFile = arg #contour.sqlite
-        elif opt in ("--intervall"): 
+        elif opt in ("--interval"): 
             interval = arg #5
         elif opt in ("--pixelSize"):
             pixelSize = arg #10
         elif opt in ("--gaussainBlur"):
             gaussainBlur = arg #{3, 5, 7, 9, 13}
+        elif opt in ("--addContoursBuffer"):
+            contourBuffer = arg #50           
     
     if (inputDEM != '') & (outputFile != '') & (interval != '') & (pixelSize != ''):
 
         # Resample
-        os.system("gdal_translate " + inputDEM + " " + tmp_file + " -tr "  + str(pixelSize) + " " + str(pixelSize) + " -r cubic")
+        os.system("gdal_translate " + inputDEM + " " + tmp_file + " -tr "  + str(pixelSize) + " " + str(pixelSize) + " -r cubic -q")
 
         # Smooth terrain
         smooth_dem = smoothTerrain(tmp_file, gaussainBlur)
@@ -234,8 +244,8 @@ def main(argv):
         if(overwrite):
             try:
                 os.remove(outputFile)
-                print("Overwrite old file.")
-                createContours(outputFile, smooth_dem, interval)
+                print("Overwrite old file...")
+                createContours(outputFile, smooth_dem, interval, contourBuffer)
             except:
                 if os.path.isfile(outputFile):
                     print("Can not delete old file! Is it open in a GIS?")
